the merged landing page and controllers of respective views
