+++ app/Http/Controllers/AboutController.php
@@ -1,0 +1,13 @@
+<?php
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+use Illuminate\Routing\Controller;
+
+class AboutController extends Controller
+{
+    public function about()
+    {
+        return view('layouts.about');
+    }
+}
