+++ app/Http/Controllers/ContactController.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+use Illuminate\Routing\Controller;
+
+class ContactController extends Controller
+{
+    public function contact()
+    {
+        return view('layouts.contact');
+    }
+}
