+++ app/Http/Controllers/HomeController.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+use Illuminate\Routing\Controller;
+
+class HomeController extends Controller
+{
+    public function welcome()
+    {
+        return view('layouts.welcome');
+    }
+}
