--- resources/views/footer.blade.php
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Document</title>
-</head>
-<body>
-    <footer class="copyright">
-        <p>Copyright &copy;Hepi Events.com. Where trust meets perfection</p>
-    </footer>
-</body>
-</html>
