+++ resources/views/layouts/about.blade.php
@@ -1,14 +1,9 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link rel="stylesheet" href="styles.css">
-    <title>about page</title>
-</head>
-<body class="about">
-<div id="about">
+@extends('layouts.merge')
+
+@section('title','About Page')
+
+@section('content')
+<section id="about">
     <div class="container">
         <div class="about-row">
             <div class="about-col-1">
@@ -39,6 +34,5 @@
             </div>
         </div>
     </div>
-</div>
-</body>
-</html>
+</section>
+@endsection
