+++ resources/views/layouts/welcome.blade.php
@@ -1,0 +1,11 @@
+@extends('layouts.merge')
+
+@section('title','Home Page')
+
+@section('content')
+<section id="header">
+    <h1>Hepi Events</h1>
+    <P>Where Trust meets Perfection</P>
+    <button type="button"  class="custom-button"><a href="/register">Get started</a></button>
+</section> 
+@endsection
