--- resources/views/news.blade.php
@@ -1,77 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link rel="stylesheet" href="styles.css">
-    <title>news page</title>
-</head>
-<body class="event-body">
-<div class="slideshow-container">
-
-    <div class="mySlides fade">
-        <img src="posters/poster1.jpg" style="width:100%" >
-    </div>
-
-    <div class="mySlides fade">
-        <img src="posters/poster2.jpg" style="width:100%">
-    </div>
-
-    <div class="mySlides fade">
-        <img src="posters/poster3.jpg" style="width:100%">
-    </div>
-
-    <div class="mySlides fade">
-        <img src="posters/poster4.jpg" style="width:100%">
-    </div>
-
-    <div class="mySlides fade">
-        <img src="posters/poster5.jpg" style="width:100%">
-    </div>
-</div>
-
-<a class="prev" onclick="plusSlides(-1)">❮</a>
-<a class="next" onclick="plusSlides(1)">❯</a>
-
-<br>
-
-<div style="text-align:center">
-    <span class="dot" onclick="currentSlide(1)"></span> 
-    <span class="dot" onclick="currentSlide(2)"></span> 
-    <span class="dot" onclick="currentSlide(3)"></span>
-    <span class="dot" onclick="currentSlide(4)"></span> 
-    <span class="dot" onclick="currentSlide(5)"></span> 
-</div>
-
-{{----------------------------------------js----------------------------------}}
-<script>
-let slideIndex = 1;
-    showSlides(slideIndex);
-    
-    function plusSlides(n) {
-        showSlides(slideIndex += n);
-    }
-    
-    function currentSlide(n) {
-        showSlides(slideIndex = n);
-    }
-    
-    function showSlides(n) {
-        let i;
-        let slides = document.getElementsByClassName("mySlides");
-        let dots = document.getElementsByClassName("dot");
-        if (n > slides.length) {slideIndex = 1}    
-        if (n < 1) {slideIndex = slides.length}
-        for (i = 0; i < slides.length; i++) {
-        slides[i].style.display = "none";  
-        }
-        for (i = 0; i < dots.length; i++) {
-            dots[i].className = dots[i].className.replace(" active", "");
-        }
-        slides[slideIndex-1].style.display = "block";  
-        dots[slideIndex-1].className += " active";
-    }
-</script>
-</body>
-</html>
