--- resources/views/welcome.blade.php
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <script src="https://kit.fontawesome.com/0538fc5c28.js" crossorigin="anonymous"></script>
-    <link rel="stylesheet" href="styles.css">
-    <title>welcome page</title>
-</head>
-<body class="back-ground">
-
-{{-------------------------------------------navbar----------------------------------}} 
-
-<div class="header">
-    <nav id="navbar">
-        <ul id="sidemenu">
-            <li class="logo"><img src="/images/logo.png"></li>
-            <li><a href="/">Home</a></li>
-            <li><a href="/about">About</a></li>
-            <li><a href="/news">News</a></li>
-            <li><a href="/contact">Contact</a></li>
-            <li><a href="/events">Events</a></li>
-            <li><a href="/admin">Admin</a></li>
-            <li><a href="/login">Login</a></li>
-        </ul>
-    </nav>
-</div>
-
-{{-------------------------------------------welcome message----------------------------}}
-<div class="header">
-    <h1>Hepi Events</h1>
-    <P>Where Trust meets Perfection</P>
-    <button type="button"  class="custom-button"><a href="/register">Get started</a></button>
-</div> 
-</body>
-</html>
