+++ routes/web.php
@@ -2,6 +2,22 @@
 
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\PageController;
+use App\Http\Controllers\HomeController;
+use App\Http\Controllers\AboutController;
+use App\Http\Controllers\NewsController;
+use App\Http\Controllers\ContactController;
+
+// Home page
+Route::get('/', [HomeController::class, 'welcome']);
+
+// About page
+Route::get('/about', [AboutController::class, 'about']);
+
+// News page
+Route::get('/news', [NewsController::class, 'news']);
+
+// Contact page
+Route::get('/contact', [ContactController::class, 'contact']);
 
 /*
 |--------------------------------------------------------------------------
@@ -13,24 +29,3 @@ use App\Http\Controllers\PageController;
 | contains the "web" middleware group. Now create something great!
 |
 */
-
-Route::get('/', function () {
-    return view('welcome');
-});
-
-Route::get('/about', function () {
-    return view('about');
-});
-
-Route::get('/contact', function () {
-    return view('contact');
-});
-
-Route::get('/news', function () {
-    return view('news');
-});
-
-Route::get('/login', function () {
-    return view('login');
-});
-
